Enforce the lint ruleset across the codebase; fix what it found
Ruff runs clean over the whole repository under the ruleset committed with
pyproject.toml (F, E, W, I, B, C4, UP, RUF at 100 columns). 1143 findings: 963
auto-fixed, the rest reviewed and resolved individually.

The bulk is mechanical -- PEP 585/604 annotations (the floor is 3.12, so
`dict[str, X]` is simply the correct spelling), import sorting, 81 genuinely
unused imports. What the linter surfaced that was worth having:

  * graph/auth_graph_anomaly.py computed `dst` from the computer name in the
    4624/4625 branch and never used it. Dead code on the detection path.
  * ingestion/source.py enumerated line numbers and threw them away, while
    IngestStats' own docstring argues that silently dropping records hides a
    data-quality problem. A count says the feed is damaged; it does not say
    where. `first_unparseable` now records "file:line" of the first bad record.
  * scripts/ablate_graph_views.py computed a `unique` column with a formula
    that reduces to len(caught), never printed it, and printed an empty field
    and a duplicate of `det` in its place. Replaced with the quantity a
    standalone ablation actually wants: attacks this view alone catches that
    the full configuration misses -- direct evidence of view displacement.
  * sysmon_generator.py classified destinations as internal/external and put
    the result nowhere; Sysmon EID 3 carries no such field.
  * `zip(out, mask)` in the FDR path is equal-length by construction, so
    strict=True turns an assumption into something Python checks.
  * `.strip('\\\\')` was a character-set strip masquerading as a prefix strip.
  * FeatureKind was `(str, Enum)`; on a 3.12 floor that is StrEnum. Safe here
    because contract_hash reads `.value`, so the pinned digest does not move.

Two suppressions, both reasoned rather than blanket. E501 is not enforced in
enterprise_simulator/ because it is a corpus of verbatim Windows artefacts --
command lines, registry paths, column-aligned inventories -- where wrapping
either falsifies the artefact or destroys the alignment that is the point of
the table; nor in graph/visualize.py, whose long lines are an embedded D3
template and are not Python. RUF002/RUF003 are off globally because the
docstrings carry mathematics and box-drawn tables that the rule reads as
ambiguous unicode; RUF001 still guards actual strings.

No formatter was run. `ruff format` would reflow ~21k lines, including
docstrings whose alignment is deliberate, for no defect found.

Verified nothing moved:
  * 158 tests pass, 2 opt-in network tests skipped
  * simulator output byte-for-byte identical to pre-lint across every channel
  * headline unchanged at 54/60 = 90.0% @ 3.19 FP/day over the same 6 seeds,
    with the same per-technique breakdown

+++ ueba-pipeline/ueba_pipeline/models/periodicity.py
@@ -60,7 +60,6 @@ from __future__ import annotations
 
 import math
 from dataclasses import dataclass
-from typing import Optional
 
 import numpy as np
 
@@ -96,7 +95,7 @@ def fisher_g_pvalue(g: float, m: int) -> float:
     g = float(min(max(g, 0.0), 1.0 - 1e-15))
     if g <= 0.0:
         return 1.0
-    b = int(math.floor(1.0 / g))
+    b = math.floor(1.0 / g)
     # Cap the number of terms: beyond this, g0 is so small the tail probability is
     # ~1 and the cost is not worth paying (identity typing only reads the small-p
     # tail). For k <= b, k*g <= 1, so every (1 - k*g) base is non-negative.
